Vector Verification: Rework to run based on env variable `DUCKDB_DEBUG_VERIFY_VECTOR` and move to `Main.yml` (https://redirect.github.com/duckdb/duckdb/pull/16957)
This PR modifies the vector verification tests so they run based on an
environment variable, e.g.:

```sql
DUCKDB_DEBUG_VERIFY_VECTOR=dictionary_expression build/debug/test/unittest ...
```

This then allows all of the vector verification tests to be run in a
single workflow instead of requiring multiple re-compilations.

We move this workflow from nightly to `Main.yml` as well.

In addition, we fix a minor bug with `Reinterpret` on dictionary vectors
which fixes an issue in the sequence operator verification run.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9c0ec2c66764d039d412878bbc4c589adec6c393