Pass down CMAKE_POLICY_VERSION_MINIMUM and fix for local development (https://redirect.github.com/duckdb/duckdb/pull/16953)
CMAKE_POLICY_VERSION_MINIMUM needs currently to be passed as env
variable, so fixing failures in Wasm and possibly elsewhere for when
CMake is bumped to 4.0.0

Also fixing a corner case, unsure if ever worked, where a slash would be
missing and `some/path` + "`./vcpkg.json` would be constructed as
`some/path./vcpkg.json` that do not work. Now `some/path/./vcpkg.json`.
With corner case for empty paths.

Upstream-commit: https://github.com/duckdb/duckdb/commit/42819dbb6727083909140efe59ae15e94dd80a21