Fix for selecting NaN values from Parquet files (https://redirect.github.com/duckdb/duckdb/pull/16962)
The statistics in Parquet files are not sufficient to know whether or
not `NaN` values are present in a file. There is work on exposing this
in https://github.com/apache/parquet-format/pull/196 (but given that
that is two years old - who knows if/when that will get merged and when
writers will start supporting it).

For now we need to assume `NaN` is present in all floating point
columns. This PR makes the following changes to facilitate this:

* Stats emitted by floating point columns are now `[min, NaN]` instead
of `[min, max]`. Because `NaN` in DuckDB is larger than all other values
- we must set it at the max value.
* For pruning, we do not use this bound - but instead search two bounds:
`[min, max]` and `[NaN, NaN]`. As a result, some scan pruning can still
happen as long as the filter does not have an unbounded upper bound
(i.e. we cannot prune anymore for `x > 5`, but we can prune for `x = 5`
or for `x > 5 and x < 10`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b310af21200edb8842341101aa9969c85c90dd72