Add a setting `scheduler_process_partial` that allows partial scheduling of tasks in the background threads (https://redirect.github.com/duckdb/duckdb/pull/16973)
This PR adds a new setting - `scheduler_process_partial` - that allows
partial scheduling of tasks in the background threads. Normally, the
background tasks pass `TaskExecutionMode::PROCESS_ALL` to the tasks they
execute - leading to them being stuck on processing the task fully until
it is completed. As a result, long-running queries can block
short-running queries.

In regular DuckDB operation - that is not a big problem, because every
connection has its own "processing thread" that works only on the
queries it issues - and hence short-running queries still make some
progress. That is not always the case - however. When using certain APIs
(e.g. the pending query API) the foreground thread might not participate
in the query processing, leaving everything to the background tasks. As
a result, short-running queries can become starved.

Enabling this setting improves fairness by only working on tasks
partially before rescheduling. As a result, bigger tasks will need more
"rounds" and short-running queries will complete without stalling. The
setting is disabled by default for now as in most instances this will
not be required - but is enabled in CI with the alternative verification
setting.

Upstream-commit: https://github.com/duckdb/duckdb/commit/a299d157417a148a88661bd843a02d57b86edc6f