Detect when tables have been dropped or altered, and prevent deletes in this scenario (https://redirect.github.com/duckdb/duckdb/pull/17018)
This PR fixes an issue in transaction management around the handling of
deletes and dropped tables. Previously we would not correctly detect
that a different transaction had dropped a table we had outstanding
deletes for. This could result in exceptions being thrown as well as
non-replayable WAL files (as the WAL would contain entries like `DROP
TABLE -> DELETE FROM TABLE`).

This PR now makes it so these scenarios are correctly detected, and a
transaction conflict is thrown if one transaction drops a table and the
other tries to delete from that same table.

Upstream-commit: https://github.com/duckdb/duckdb/commit/641076b4d5d21f5e6b92c6aa7a1d8aa3706c34ee