MultiFileReader Rework (part 18): Replace file path with `OpenFileInfo` struct (https://redirect.github.com/duckdb/duckdb/pull/17071)
This PR reworks several parts of the file system and multi file reader
to work on an extensible `OpenFileInfo` struct instead of a `string
path`:

* `OpenFileExtended` takes an `OpenFileInfo`
* `Glob` returns a list of `OpenFileInfo`
* Multi file readers operate on `OpenFileInfo` (including the
`MultiFileList` that returns a list of `OpenFileInfo`)


The idea here is that we can pass per-file information down to the
readers and file system that can avoid doing unnecessary work. For
example:

* When performing an S3 glob, the result set contains information about
each file (file size, eTag and last modified time). When reading the
file, we then perform a head request to obtain this information again.
This is an unnecessary round-trip per-file as we already have this
information. Using the new infrastructure we can extend the
Glob/OpenFileExtended to avoid this round-trip.
* When reading from Lakehouse formats, they generally store the file and
footer size for the data files - again allowing us to avoid the
subsequent head request and avoiding a round trip if we can pass down
this information.

This PR does not change anything aside from wrapping the paths into a
struct - but that will allow these optimizations in a follow-up PR.

Upstream-commit: https://github.com/duckdb/duckdb/commit/c69efd5a310c6a0cee2bb0235dd7c53e690e2204