Grow string dictionary dynamically in Parquet writer (https://redirect.github.com/duckdb/duckdb/pull/17061)
The `PrimitiveDictionary` that we've added for dictionary compression in
Parquet (only on `main` now, and will be released in v1.3.0) would
statically allocate `STRING_DICTIONARY_PAGE_SIZE_LIMIT` bytes for the
strings. It defaults to 1 MB. If the total string size exceeds this,
DuckDB bails on dictionary compression.

This is a bit restrictive, and does not work well when there are many
long strings that aren't unique. This PR makes it so that we initialize
at 1 MB, and then we dynamically grow until
`STRING_DICTIONARY_PAGE_SIZE_LIMIT`, which now defaults to 1 GB. This
allows large strings to still be dictionary-compressed by our writer.

Upstream-commit: https://github.com/duckdb/duckdb/commit/5538f966a910f51e3ae1146c0537f6b0d78d955f