Add extended file info to OpenFileInfo, and use this to pass encryption keys and footer size to Parquet reader (https://redirect.github.com/duckdb/duckdb/pull/17085)
Follow-up to https://github.com/duckdb/duckdb/pull/17071

This PR extends the  `OpenFileInfo` struct with `ExtendedOpenFileInfo`:

```cpp
struct ExtendedOpenFileInfo {
	unordered_map<string, Value> options;
};
```

The Parquet reader is extended to support two keys in this struct:

* `encryption_key`: used to pass a global encryption key for the Parquet
file to the reader
* `footer_size`: used to pass the footer size of the Parquet file to the
reader. This can be used to skip the separate reading of the footer
length from the back of the file - instead we can directly read the
footer. Note that this must be exactly correct - an error is thrown if
an incorrect footer size is provided.

Upstream-commit: https://github.com/duckdb/duckdb/commit/b19550535b4900232a1cd34308417f9bc868da4a