Changing mbedtls encryption API (https://redirect.github.com/duckdb/duckdb/pull/16196)
This PR replaces the mbedtls gcm encryption API (`mbedtls_gcm_xxx`) to
the general mbedtls encryption API (`mbedtls_cipher_xxxx`). It also
patches the httpfs extension to be compatible with the new function
signature of the encryption state.
 
The main reason for replacing the old API with the new one is to be able
to support `AES_GCM_CTR_V1` for Parquet encryption.  I thought it might
be better to split this up in two PR's.
 
The change of the encryption API is beneficial such that (i) the mbedtls
[en/de]cryption flow is now exactly similar to OpenSSL leading to a
single decryption flow in Parquet and (ii) the new API can handle _any_
cipher supported by mbedtls, which is e.g. necessary for supporting
`AES_GCM_CTR_V1` for Parquet encryption.
 
Note that Windows Extensions is failing in the CI due to the httpfs
patch. I have opened an Issue for this (https://redirect.github.com/duckdb/duckdb/pull/16177).

Upstream-commit: https://github.com/duckdb/duckdb/commit/93a9d9f9b58341c3d8d72c259cddb9696a77468a