Fix an issue in upserts where the local append state was not correctly flushed (https://redirect.github.com/duckdb/duckdb/pull/17109)
This PR fixes an issue in the upsert code where the local append state
was not correctly flushed. The upsert code would try to keep the local
append state active while performing the upsert - but this added a lot
of complication. This PR simplifies the code by removing the management
of the local append state and instead running the `Initialize -> Append
-> Finalize` flow on every append. This fixes a remaining issue around
incorrect flushing of the append state in the upsert.

Upstream-commit: https://github.com/duckdb/duckdb/commit/892a03d26e5f887f63e86259257f11bc50a32269