[Dev] Add "registries" to `vcpkg.json`, add script to list the packages of the registry. (https://redirect.github.com/duckdb/duckdb/pull/17124)
This PR adds <https://github.com/duckdb/vcpkg-duckdb-ports> to the
generated vcpkg.json file.
This serves as an alternative to the copy-and-pasted ports we had
previously.

All of the ports currently in (out of tree) extensions should move to
the registry, this avoids needing to copy and paste the port when
another extension depends on the same dependency, or even depends on a
different extension that has a port.

Upstream-commit: https://github.com/duckdb/duckdb/commit/f2de8fd02efbf12081e3dd229f3fa512782724a4