External File Cache (https://redirect.github.com/duckdb/duckdb/pull/16463)
This PR implements an in-memory cache for external files (e.g.,
Parquet/CSV/JSON) that caches reads in memory to speed up subsequent
reads. It is enabled by default, but can be disabled with `SET
enable_external_file_cache=false;`.

## `CachingFileSystem`
The main implementation is in `CachingFileSystem`, a class that
(sort-of) wraps `FileSystem`, but with a slightly different API. Instead
reading a byte range from a file like so:
```c++
void Read(void *buffer, idx_t nr_bytes, idx_t location);
```
Which reads `nr_bytes` from `location` _into_ `bufffer`, which must be
pre-allocated to fit `nr_bytes`, we instead read like so:
```c++
BufferHandle Read(data_ptr_t &buffer, idx_t nr_bytes, idx_t location);
```
Which returns a `BufferHandle` to a managed buffer holding the requested
`nr_bytes` from `location`, and `buffer` is set to point into it. This
entails that we can (in theory) do zero-copy reads if the data is
already cached in memory.

This is currently integrated into the Parquet reader, but only when
prefetching is enabled (which is the default remote reads, but can be
enabled for local Parquet files with `SET
prefetch_all_parquet_files=true;`).

Reads have a size of `nr_bytes`, and are identified by `path` and
`location`. If the same byte range is (partially) queried twice, we
first try to read from the in-memory cache. If a larger byte range fully
encompasses a smaller byte range (e.g., when a full row group is queried
after first querying a single column), the smaller byte range is removed
in favor of the larger byte range. The smaller byte range is used to
construct the larger byte range.

## Cache Invalidation
Because the buffers holding the byte ranges are in-memory only (for
now), they might become invalid if the buffer manager decides to evict
them. If an invalid buffer is encountered when trying to read from the
corresponding file, it is removed from the cache.

Another reason for cache entries becoming invalid is due to files being
modified. If we detect that a file has been modified, all of the byte
ranges of that file are removed from the cache. We detect this for
remote files using the HTTP `Etag` field.

For local files, we use the "last modified" timestamp. For some file
systems, the resolution of this timestamp is in nanoseconds, which is
great, since reads/writes are so fast nowadays, we really need that to
be sure that files have not been modified in the meantime. However, some
file systems use a resolution of _seconds_, and our implementation
should work regardless of the underlying file system and hardware. So,
we require the last modified time to be more than _two seconds_ in the
past for the cache entry to be considered valid.

## Example: Remote File
The most compelling example is for remote reads. For example:
```sql
D .timer on
D .mode trash
D from 's3://duckdb-blobs/data/shakespeare.parquet';
Run Time (s): real 1.826 user 0.042848 sys 0.024513
D from 's3://duckdb-blobs/data/shakespeare.parquet';
Run Time (s): real 0.661 user 0.043247 sys 0.005002
```
As we can see, the second read is much faster than the initial read. In
theory, it could be just a few milliseconds, but the second query still
checks the `Etag` to see if the file was modified in between queries. We
may add a config setting to skip checking this, but this can of course
lead to very confusing errors when parts of the file are cached, and
other are not, but the file has been changed in between queries.

The external file cache can be queried to check which files and byte
ranges are currently in the cache:
```sql
D .mode duckbox
D from duckdb_external_file_cache();
┌────────────────────────────────────────────┬──────────┬──────────┬─────────┐
│                    path                    │ nr_bytes │ location │ loaded  │
│                  varchar                   │  int64   │  int64   │ boolean │
├────────────────────────────────────────────┼──────────┼──────────┼─────────┤
│ s3://duckdb-blobs/data/shakespeare.parquet │  1697483 │        4 │ true    │
│ s3://duckdb-blobs/data/shakespeare.parquet │      697 │  1697487 │ true    │
└────────────────────────────────────────────┴──────────┴──────────┴─────────┘
```
As we can see, there are two byte ranges in cache: a small one for the
footer, and a larger one for the single row group that is in the file.

External file cache entries also show up under their own tag in
`duckdb_memory()`:
```sql
D from duckdb_memory() where tag = 'EXTERNAL_FILE_CACHE';
┌─────────────────────┬────────────────────┬─────────────────────────┐
│         tag         │ memory_usage_bytes │ temporary_storage_bytes │
│       varchar       │       int64        │          int64          │
├─────────────────────┼────────────────────┼─────────────────────────┤
│ EXTERNAL_FILE_CACHE │      1703936       │            0            │
└─────────────────────┴────────────────────┴─────────────────────────┘
```

## Example: Local File
I have the TPC-H `lineitem` table at SF10 in a local Parquet file
(`lineitem10.parquet`).
When I run the following query on current `main`
```sql
select any_value(columns(*)) from 'lineitem10.parquet';
```
I get an execution time of ~0.44s, and `SET
prefetch_all_parquet_files=true;` does not seem to make much of a
difference.

On this branch, with `SET prefetch_all_parquet_files=true;`, I get an
execution time of ~0.46s on the first read, and an execution time of
~0.39s on subsequent reads.

So, there seems to be a slight overhead from enabling the cache on the
first read (most likely due to all threads inserting/reading from the
cache in parallel, requiring some amount of locking), but subsequent
reads are a bit faster.

## Future Work
In the future, I'd like to enable this not only for prefetch reads, but
for _almost all_ Parquet reads. Currently, this would cause overhead,
since the Parquet reader copies the data from the cached buffer into a
local buffer. However, this can (and should) be entirely zero-copy. We
can also cache decompressed buffers, which would speed up subsequent
reads significantly.

We could also enable persistent caching, so that users can close their
remote datalake querying session, and pick it back up later without such
a high initial cost of reading files over the network.

Finally, we could consider integrating this into the CSV/JSON reader as
well, so that users can enjoy the same benefits for those file formats
as well.

Lots of work to do still, but I think it's best to send this PR now to
receive some feedback before going all-in on all kinds of optimizations.

EDIT: I've made it so we also cache the last 8 bytes of Parquet files
(footer_length + "PAR1"), so this only opens the file handle when
queried for the second time, all else is cached, reducing the time of
the second read of the remote Parquet file down from 0.78s to 0.66s.

Upstream-commit: https://github.com/duckdb/duckdb/commit/9e7fedd4c0af54f16f8fd89367f4cdb87295d22e