Notify nightly build status (https://redirect.github.com/duckdb/duckdb/pull/17108)
The idea is to trigger the `NightlyBuildsStatus.yml` in the end of the
`InvokeCI.yml` run.
There is a `NotifyExternalRepositories.yml` workflow for triggering
workflow run on the external repos - we just need to pass there all the
inputs needed by `NightlyBuildsStatus.yml` run (branch name, event name,
should the status report be published or not).

The `NotifyExternalRepositories.yml` triggers two workflows:
[duckdb-odbc](https://github.com/duckdb/duckdb-odbc/tree/main)/[.github](https://github.com/duckdb/duckdb-odbc/tree/main/.github)/[workflows](https://github.com/duckdb/duckdb-odbc/tree/main/.github/workflows)/`Vendor.yml`
and
[duckdb-build-status](https://github.com/duckdb/duckdb-build-status/tree/main)/[.github](https://github.com/duckdb/duckdb-build-status/tree/main/.github)/[workflows](https://github.com/duckdb/duckdb-build-status/tree/main/.github/workflows)`/NightlyBuildsCheck.yml`
workflows.
`Vendor.yml` should be triggered only when all builds had succeeded.
`NightlyBuildsStatus.yml` should run in any case. 

So we need to add one more job to `InvokeCI.yml` to create `is-success`
status, which is `true` only when **all** builds had succeeded. Then
`is-success` status should be passed to the
`NotifyExternalRepositories.yml` and its value should be checked before
the `Vendor.yml` workflow run is called. We should add `always()`
condition to the "notifying" jobs `if:` conditions - that makes them not
get skipped in case there a builds not succeeded ([tested
here](https://github.com/hmeriann/trigger_nightly_check/actions/runs/14446186717)).

Upstream-commit: https://github.com/duckdb/duckdb/commit/ec2173c65b5be1737eeda87e5d0f935ab8e4739b