[ENCRYPTION] Make block header size adaptive (https://redirect.github.com/duckdb/duckdb/pull/17118)
This PR adds the possibility to change the block header size, which is
necessary for block-based database file encryption. Before this, the
header of a persistent block could only consist of an 8-byte checksum.
For the encryption of blocks, we need to store some metadata (typically
a 12-byte nonce and a 16-byte tag) at the start of each block _before_
the computed checksum, since the checksum will also be encrypted.

For now, we fix the the `ENCRYPTED_BLOCK_HEADER_SIZE` to 40 bytes, which
includes the 8-byte checksum. This will be enough for the default
encryption algorithm (`AES-GCM`), which requires a 12-byte nonce and a
16-byte tag. The block header size needs to be 8-byte aligned, because
the checksum is calculated in blocks of 64-bits.

Changing the block header size should _only_ happen when a reading or
writing an encrypted database file. To trigger this, I've added some
logic to add an (empty) encryption key. Just adding `ENCRYPTION_KEY` at
attach will force the block header size to change, but note that there
is no actual encryption yet. Trying to read a database file with a
different block header size will now throw an invalid checksum error. In
a follow up PR I will add proper encryption key logic, but the
encryption flow (`ATTACH "encrypted_db.duckdb" as encrypted
(ENCRYPTION_KEY "asdf")`) will be similar.

Note that the block header of the main database header will remain 8
bytes. This is because the main header will _not_ be encrypted since (i)
it always needs to be readable and (ii) the main header typically does
not store any sensitive data. This does add somewhat more complicated
logic in the `ChecksumAndRead()` and `ChecksumAndWrite()` methods, which
is mainly because the block manager is constructed before the file
buffer is created.

Edit:

I've added some new tests + logic to properly handle the checksum
errors. Now:

- An encrypted file is rejected when read by a duckdb <  v1.3.0
- A clean error is thrown if the database file is encrypted, but no
`ENCRYPTION_KEY` is specified
- A clean error is thrown if an `ENCRYPTION_KEY` is specified, but the
database exists and is NOT encrypted

Upstream-commit: https://github.com/duckdb/duckdb/commit/c0e88f8cc6e8007d2fef5cadb7ca59aa712c9d55