Add `SYSTEM_PEAK_BUFFER_MANAGER_MEMORY` and `SYSTEM_PEAK_TEMP_DIRECTORY_SIZE` to profiler (https://redirect.github.com/duckdb/duckdb/pull/17164)
Measures the peak buffer manager memory usage and temp directory size
_of the entire system_ during the query. We could consider expanding
this to more query-specific metrics in the future, but these numbers are
already easily available; this just exposes them.

They are not enabled by default, but maybe they should be? I would be
happy to receive feedback on this.

Upstream-commit: https://github.com/duckdb/duckdb/commit/8a17f332fe147a416dff32c10bc157909fdf29d4