Fix [InvokeCI / NotifyExternalRepository] Unexpected value 'true' (https://redirect.github.com/duckdb/duckdb/pull/17212)
[InvokeCI](https://github.com/duckdb/duckdb/actions/runs/14584113147/job/40919219118)
keeps skipping the external repos notification, which is more likely
because of not matching types of the `should-publish.

The `.github/workflows/NotifyExternalRepositories.yml` was expecting
`boolean` values passed as the inputs from the
`.github/workflows/InvokeCI.yml`, but the `prepare-status` job set a
string value into a `is-success` GH outputs variable.

Also we don't want to notify the Vendor when the `InvokeCI` hadn't
succeeded, so the `.github/workflows/NotifyExternalRepositories.yml`
should only trigger the `Vendor.yml` run on `if: ${{ inputs.is-success
== 'true' }}`

This PR fixes mentioned problems and corrects the inputs names passed to
the `NightlyBuildCheck.yml`.

Upstream-commit: https://github.com/duckdb/duckdb/commit/de576b8f23862a9bff56ac8666816ec1df339a7d