Add support for the cast_to_type function, that allows generating a cast from an expression to the type of another column (https://redirect.github.com/duckdb/duckdb/pull/17209)
This PR adds the `cast_to_type` function. This function is transformed
into a cast from the first input argument to the type of the second
input argument. For example:

```sql
SELECT cast_to_type('42', NULL::INT) AS res;
┌───────┐
│  res  │
│ int32 │
├───────┤
│  42   │
└───────┘
```

This function is primarily useful in macros, as it allows us to maintain
types. This helps with making generic macros (that operate on different
types). For example, here is a macro that adds to a number if the input
is an integer:

```sql
CREATE TABLE tbl(i INT, s VARCHAR);
INSERT INTO tbl VALUES (42, 'hello world');

CREATE MACRO conditional_add(col, nr) AS
    CASE WHEN typeof(col) == 'INTEGER'
    THEN cast_to_type(col::INTEGER + nr, col)
    ELSE col END;
SELECT conditional_add(COLUMNS(*), 100) FROM tbl;
┌───────┬─────────────┐
│   i   │      s      │
│ int32 │   varchar   │
├───────┼─────────────┤
│  142  │ hello world │
└───────┴─────────────┘
```

The way this works is that the `CASE` statement needs to return the same
type in all code paths. We can perform the addition on any input column
by adding a cast to the desired type - but we need to cast the result of
the addition back to the source type to make the binding work. Without
the cast we end up with the following error message:

```sql

CREATE MACRO broken_conditional_add(col, nr) AS
    CASE WHEN typeof(col) == 'INTEGER'
    THEN col::INTEGER + nr
    ELSE col END;
SELECT broken_conditional_add(COLUMNS(*), 100) FROM tbl;
-- Cannot mix values of type VARCHAR and INTEGER in CASE expression - an explicit cast is required
```

So we need a way of dynamically casting to the type of a column - which
this PR adds using the `cast_to_type` function.

Upstream-commit: https://github.com/duckdb/duckdb/commit/db55c74c759a344d837583884058c5ad21455358