Add `ExternalFileCache` validation as option for `ExtendedOpenFileInfo` (https://redirect.github.com/duckdb/duckdb/pull/17205)
This PR adds functionality to bypass the built-in cache validation for
`ExternalFileCache` by adding the key `"validate_external_file_cache"`
with value `false`, to the `ExtendedOpenFileInfo` of `OpenFileInfo`.
This is not exposed to users, so it can only be used internally. This
avoids doing a `HEAD` request for files that have been queried before.

The Iceberg/Delta extensions should be able to set this to `false`
safely. cc @Tmonster

Upstream-commit: https://github.com/duckdb/duckdb/commit/84db85c090101b3c420b939bfe80b93ae8f4d771