[Dev] Re-enable Iceberg, Bump Avro, fix `generate_extension_functions.py` for dependencies between extensions (https://redirect.github.com/duckdb/duckdb/pull/17204)
This PR is a follow-up to https://redirect.github.com/duckdb/duckdb/pull/17183 

To elaborate on that last part:
In `add_extension` we first check if there are dependencies (hardcoded
in the `EXTENSION_DEPENDENCIES` dictionary)
if there are, we make sure to call `add_extension` for those first, if
they are not in the `added_extensions` set already.

In the various `add_<catalog entry type>` methods, we previously only
subtracted the `base_<catalog entry type>` set.
Now we check the dependencies and also subtract the entries added by the
dependencies of the extension, tracked by the `extension_<catalog entry
type>` dictionary.

This is also where we update the `extension_<catalog entry type>`
dictionary.
At the end of `add_extension` we add to the `added_extensions` set.

Upstream-commit: https://github.com/duckdb/duckdb/commit/ad512b91876411ee2fdae9ee6d288d86aef5fbb2