Fix [InvokeCI / NotifyExternalRepository] GitHub Actions has encountered an internal error when running your job. (https://redirect.github.com/duckdb/duckdb/pull/17218)
GitHub Actions didn't provide more details in the [error
message](https://github.com/duckdb/duckdb/actions/runs/14607342802/job/40992045763),
so I ran some test workflows.

We are not expecting the `InvokeCI / NotifyExternalRepository` would
skip `Run ODBC Vendor` job and fail `Run JDBC Vendor` and `Run Nightly
Build Status` jobs.

The issue was in the boolean values passed as string inputs to the
called `NotifyExternalRepository` workflow, where they were interpreted
as booleans
([here](https://github.com/hmeriann/duckdb/blob/main/.github/workflows/NotifyExternalRepositories.yml#L60)).

This PR fixes that and adds missing `if` conditions to both workflows:
to the `prepare-status:` job in `InvokeCI` (that job creates a flag to
trigger or not the ODBC/JDBC Vendors) and to the `notify-jdbc-run` job
in `NotifyExternalRepositories` (to make it have the same behaviour as
`notify-odbc-run`)

Upstream-commit: https://github.com/duckdb/duckdb/commit/4ba87eb4d09c1a827085b76c4e467508c5d53b3e