Only notify JDBC when all runs are successful (https://redirect.github.com/duckdb/duckdb/pull/17233)
This change adds the check to JDBC notification workflow that all
nightly CI runs are sucessful. The check is added the same way it is
used for ODBC notifications.

This check is necessary to make sure, that JDBC/ODBC only imports engine
sources for the commit for which the whole set of extensions is built
and available.

Upstream-commit: https://github.com/duckdb/duckdb/commit/8979b173f7d591b20beb53c4df208b8266476696