Dont try to load extension if storage type is already registered (https://redirect.github.com/duckdb/duckdb/pull/17241)
This fixes an issue that occurs when an extension provides a storage
type that is not named the same as the extension.

Even if the extension was loaded, we would still fall into the
`ExtensionHelper::LoadExternalExtension` path and try to load an
extension with the name of the storage type, throwing an error. Now we
first check if a storage type with the target name has already been made
available by a previously loaded extension.

Upstream-commit: https://github.com/duckdb/duckdb/commit/cf09421cb5f03a3ab6984e0011d6c6214070ab16