Guess Parquet footer size (https://redirect.github.com/duckdb/duckdb/pull/17300)
... based on file size, when the file is remote.

Round trips to a remote file source are very expensive. When reading a
Parquet file, we currently need 3 round trips to get the footer (one
`HEAD`, one `GET` to read the footer size, and another `GET` to read the
footer given that size). This PR implements guessing the footer size. If
we're wrong, we end up with 3 round trips like before. If we're right,
we only need 2 round trips.

We guess the Parquet footer size to be 1/1000th of the file size, and
always between 16-256kb. We bound the guess to reduce the risk of
causing unnecessary overhead. At these read sizes, there is virtually no
noticeable penalty for guessing wrong, while potentially yielding a
significant benefit.

The cost is incurred once per Parquet file. When reading many Parquet
files at once with a glob pattern, this adds up. Typically, with such
queries, the Parquet files are quite small, so we're targeting these
kind of queries with this optimization. When reading large Parquet
files, our guess might be wrong, but that's (probably) OK, since the
extra round trip will likely be insignificant compared to the cost of
reading that large Parquet file.

I've also taken the liberty to fix some of our `NightlyTest` CI run
failures. I've added `set enable_external_file_cache=false;` to a few
tests so that we get the expected number of `GET` requests again, and
I've prevented some assertion failures in `CompressedMaterialization`
that were caused by our dictionary expression executor by adding
`FunctionErrors::CAN_THROW_RUNTIME_ERROR` when assertions are enabled.

Upstream-commit: https://github.com/duckdb/duckdb/commit/01f1bec790dba5f410859a0e0b09144011dbfe5e