Generating random data for mbedtls without key (https://redirect.github.com/duckdb/duckdb/pull/17309)
For generating random data with mbedtls we don't actually need to
initialise an AES state, which means we also do not need a key. However,
now, we require to initialise an encryption state, and if no proper key
is given this throws an error.

This PR fixes this behaviour by making a static generate random data
function that can be called independently of the mbedtls AES state.

Upstream-commit: https://github.com/duckdb/duckdb/commit/37ead0e18cc61dcf7bc941595e31f0d95bef6a41