Implement ARTMerger replacing the recursive ART merge algorithm (https://redirect.github.com/duckdb/duckdb/pull/17243)
This PR introduces a new class, `ARTMerger`. It takes two nodes and
merges their (sub)trees.

Currently, we merge indexes using a recursive strategy. The changes in
this PR replace this approach with an iterative merge strategy. The
`ARTMerger` also encapsulates much of the merge code into its private
functions and uses an arena allocator for many of its allocations.

Benchmark timings are similar.

Upstream-commit: https://github.com/duckdb/duckdb/commit/a2caf3038bc61be6642229877837d7e3a1b47264