[Nested] Fix: 16489 - Find `NULL`s in lists using `list_position` (https://redirect.github.com/duckdb/duckdb/pull/17080)
Previously, when calling `list_position([1, 2, 3, NULL], NULL)`, the
function would always return `NULL` instead of the index of the `NULL`
element. This behavior differed from PostgreSQL. This PR fixes the
issue, so that `list_position` now correctly returns the index of the
element, regardless of whether or not it is `NULL`.

This PR builds on: https://github.com/duckdb/duckdb/pull/16969

Fixes: https://github.com/duckdb/duckdb/issues/16489
Upstream-commit: https://github.com/duckdb/duckdb/commit/541d3fe2e693a059a9bbfb538fe01381ac079e61