fix https://redirect.github.com/duckdb/duckdb/pull/17258: Allow to open database in readonly mode within cli (https://redirect.github.com/duckdb/duckdb/pull/17375)
Since the iName variable declared before the for loop was shadowed by a
new one inside the loop, its value remained unchanged after the loop. As
a result, when it was later used to retrieve the database name, it still
held the initial value 1, which pointed to the --readonly parameter.

Upstream-commit: https://github.com/duckdb/duckdb/commit/79a639fdf7b09484441cdac3e12c646ac794d90c