Deprecate lambda arrow (->) and replace it with LAMBDA x : x + 1 (https://redirect.github.com/duckdb/duckdb/pull/17235)
## Changes

Currently, one way to specify lambda functions is by using the arrow
operator syntax: `x -> x + 1`

This PR deprecates that syntax and replaces it with Python-style lambda
syntax: `lambda x: x + 1`

You can enable the old behavior with: `SET
allow_deprecated_lambda_syntax=true` (default value is also `true`).
On default, it is currently still possible to use both the old and the
new syntax.

### Timeline

Starting with this PR / 1.3.0, we introduce the new syntax, and the
`allow_deprecated_lambda_syntax` setting.

Starting with 1.4.0, we intend to print some kind of warning when using
the old behavior and `allow_deprecated_lambda_syntax` is on the default
setting.

Starting with 1.5.0, we intend to default
`allow_deprecated_lambda_syntax` to `no allowed`.

Starting with 1.6.0, we'll disable using the deprecated lambda syntax,
and remove the setting.

### Motivation

The same arrow operator is also used in the JSON extension. The decision
to deprecate it for lambda functions stems from many issues that arise
from using the same operator for both cases. Binding this operator is a
lot of trouble, as we must resolve ambiguity conflicts and precedence
issues. The operator precedence that is necessary for the lambda arrow
to work as expected causes weird behavior for the JSON arrow operator
(`->` and `->>`).

Since Postgres uses the arrow operator for JSON operations, we modified
the lambda functions so that the arrow operator can no longer be used to
reconcile this problem.

Upstream-commit: https://github.com/duckdb/duckdb/commit/95b70e27491079c07ce4c48f7e97850781079554