Function chaining: report missing column instead of missing function if function exists (https://redirect.github.com/duckdb/duckdb/pull/17383)
When doing column function chaining error messages are confusing when
the wrong column name is used:

```sql
D select l_coment.replace('a', 'b') from lineitem;
-- Catalog Error:
-- Scalar Function with name "replace" is not in the catalog, but it exists in the core_functions extension.
```

Clearly the `replace` function exists - the error message the user is
looking for is reporting that there is a typo in `l_comment`.

This PR resolves this by checking if the function exists - and if it
exists, we throw the missing column error message instead.

Upstream-commit: https://github.com/duckdb/duckdb/commit/576e2d74e33b94bc44672814a64a5089ddf891c5